--- DBconfig.json
@@ -1,7 +1,0 @@
-{
-    "dbpath" : "./storage",
-    "pagesize" : "30",
-    "dm_maxfilesize" : "120",
-    "bm_buffercount" : "3",
-    "bm_policy" : ["LRU", "MRU"]
-}
