+++ src/config/DBconfig.json
@@ -1,0 +1,7 @@
+{
+    "dbpath" : "./storage",
+    "pagesize" : "30",
+    "dm_maxfilesize" : "120",
+    "bm_buffercount" : "3",
+    "bm_policy" : ["LRU", "MRU"]
+}
